Refactor steps
