+++ tests/e2e/features/remote_connection_close.feature
@@ -1,11 +1,10 @@
 Feature: Remote Connection closing
-  # FIXME
-#  Scenario: Connection.on_close fires on partner Member leave
-#    Given room with joined members Alice and Bob
-#    When `Bob`'s Room closed by client
-#    Then `Alice`'s Connection with `Bob` closes
+  Scenario: Connection.on_close fires on partner Member leave
+    Given room with joined members Alice and Bob
+    When Bob's room closed by client
+    Then Alice's connection with Bob closes
 
   Scenario: Connection closes on partner Member delete by Control API
     Given room with joined members Alice and Bob
-    When Control API removes Member `Bob`
-    Then `Alice`'s Connection with `Bob` closes
+    When Control API removes member Bob
+    Then Alice's connection with Bob closes
