+++ tests/e2e/steps/mod.rs
@@ -1,2 +1,5 @@
 mod connection;
+mod control_api;
 mod media_state;
+mod room;
+mod track;
