--- tests/e2e/then.rs
@@ -1,254 +1,0 @@
-use cucumber_rust::then;
-
-use crate::{
-    object::{room::FailedParsing, MediaKind, MediaSourceKind},
-    world::World,
-};
-
-#[then(regex = "^`(.*)` receives Connection with Member `(.*)`$")]
-async fn then_member_receives_connection(
-    world: &mut World,
-    id: String,
-    partner_id: String,
-) {
-    let member = world.get_member(&id).unwrap();
-    member
-        .connections()
-        .wait_for_connection(partner_id.clone())
-        .await
-        .unwrap();
-}
-
-#[then(regex = "^`(.*)` doesn't receives Connection with Member `(.*)`")]
-async fn then_member_doesnt_receives_connection(
-    world: &mut World,
-    id: String,
-    partner_id: String,
-) {
-    let member = world.get_member(&id).unwrap();
-    assert!(member
-        .connections()
-        .get(partner_id)
-        .await
-        .unwrap()
-        .is_none())
-}
-/// Parses [`MediaKind`] and [`MediaSourceKind`] from the provided [`str`].
-fn parse_media_kinds(
-    s: &str,
-) -> Result<(MediaKind, MediaSourceKind), FailedParsing> {
-    let media_kind = s.parse()?;
-    let source_kind = match media_kind {
-        MediaKind::Audio => MediaSourceKind::Device,
-        MediaKind::Video => s.parse()?,
-    };
-    Ok((media_kind, source_kind))
-}
-
-#[then(regex = "^`(.*)`'s (audio|(?:display|device) video) RemoteMediaTrack \
-                with `(.*)` is (enabled|disabled)$")]
-async fn then_remote_media_track(
-    world: &mut World,
-    id: String,
-    kind: String,
-    partner_id: String,
-    state: String,
-) {
-    let member = world.get_member(&id).unwrap();
-    let partner_connection = member
-        .connections()
-        .wait_for_connection(partner_id)
-        .await
-        .unwrap();
-    let tracks_with_partner = partner_connection.tracks_store().await.unwrap();
-
-    let (media_kind, source_kind) = parse_media_kinds(&kind).unwrap();
-    let track = tracks_with_partner
-        .get_track(media_kind, source_kind)
-        .await
-        .unwrap();
-
-    match state.as_str() {
-        "enabled" => track.wait_for_enabled().await.unwrap(),
-        "disabled" => track.wait_for_disabled().await.unwrap(),
-        _ => unreachable!(),
-    };
-}
-
-#[then(regex = "^`(.*)` doesn't have (audio|(?:device|display) video) \
-                RemoteMediaTrack with `(.*)`$")]
-async fn then_doesnt_have_remote_track(
-    world: &mut World,
-    id: String,
-    kind: String,
-    partner_id: String,
-) {
-    let member = world.get_member(&id).unwrap();
-    let partner_connection = member
-        .connections()
-        .wait_for_connection(partner_id)
-        .await
-        .unwrap();
-    let tracks_with_partner = partner_connection.tracks_store().await.unwrap();
-    let (media_kind, source_kind) = parse_media_kinds(&kind).unwrap();
-
-    assert!(!tracks_with_partner
-        .has_track(media_kind, Some(source_kind))
-        .await
-        .unwrap());
-}
-
-#[then(regex = "^`(.*)`'s Room.on_close callback fires with `(.*)` reason$")]
-async fn then_on_close_fires(
-    world: &mut World,
-    id: String,
-    expect_reason: String,
-) {
-    let reason = world.wait_for_on_close(&id).await.unwrap();
-    assert_eq!(expect_reason, reason);
-}
-
-#[then(regex = "^`(.*)`'s Connection with `(.*)` closes$")]
-async fn then_connection_closes(
-    world: &mut World,
-    id: String,
-    partner_id: String,
-) {
-    let member = world.get_member(&id).unwrap();
-    let connection =
-        member.connections().get(partner_id).await.unwrap().unwrap();
-    connection.wait_for_close().await.unwrap();
-}
-
-#[then(regex = "^Member `(.*)` has (\\d*) local Tracks$")]
-async fn then_member_has_local_tracks(
-    world: &mut World,
-    id: String,
-    count: u64,
-) {
-    let member = world.get_member(&id).unwrap();
-    let room = member.room();
-    let tracks = room.local_tracks().await.unwrap();
-    assert_eq!(count, tracks.count().await.unwrap());
-}
-
-#[then(regex = "^`(.*)` has local (audio|(?:device |display )?video)$")]
-async fn then_has_local_track(world: &mut World, id: String, kind: String) {
-    let member = world.get_member(&id).unwrap();
-    let room = member.room();
-    let tracks = room.local_tracks().await.unwrap();
-    let media_kind = kind.parse().unwrap();
-    let source_kind = kind.parse().ok();
-
-    assert!(tracks.has_track(media_kind, source_kind).await.unwrap())
-}
-
-#[then(regex = "^`(.*)` remote (audio|(?:device|display) video) Track from \
-                `(.*)` disables")]
-async fn then_remote_track_stops(
-    world: &mut World,
-    id: String,
-    kind: String,
-    remote_id: String,
-) {
-    let member = world.get_member(&id).unwrap();
-    let (media_kind, source_kind) = parse_media_kinds(&kind).unwrap();
-
-    let conn = member.connections().get(remote_id).await.unwrap().unwrap();
-    let track = conn
-        .tracks_store()
-        .await
-        .unwrap()
-        .get_track(media_kind, source_kind)
-        .await
-        .unwrap();
-    assert!(track.muted().await.unwrap());
-}
-
-#[then(regex = "^on_disabled callback fires (\\d*) time on `(.*)`'s remote \
-                (audio|(?:device|display) video) Track from `(.*)`$")]
-async fn then_on_remote_disabled_callback_fires(
-    world: &mut World,
-    times: u64,
-    id: String,
-    kind: String,
-    remote_id: String,
-) {
-    let member = world.get_member(&id).unwrap();
-    let remote_conn =
-        member.connections().get(remote_id).await.unwrap().unwrap();
-    let (media_kind, source_kind) = parse_media_kinds(&kind).unwrap();
-
-    let track = remote_conn
-        .tracks_store()
-        .await
-        .unwrap()
-        .get_track(media_kind, source_kind)
-        .await
-        .unwrap();
-    track.wait_for_on_disabled_fire_count(times).await.unwrap();
-}
-
-#[then(regex = "^on_enabled callback fires (\\d*) time on `(.*)`'s remote \
-                (audio|(?:device|display) video) Track from `(.*)`$")]
-async fn then_on_remote_enabled_callback_fires(
-    world: &mut World,
-    times: u64,
-    id: String,
-    kind: String,
-    remote_id: String,
-) {
-    let member = world.get_member(&id).unwrap();
-    let remote_conn =
-        member.connections().get(remote_id).await.unwrap().unwrap();
-    let (media_kind, source_kind) = parse_media_kinds(&kind).unwrap();
-    let track = remote_conn
-        .tracks_store()
-        .await
-        .unwrap()
-        .get_track(media_kind, source_kind)
-        .await
-        .unwrap();
-    track.wait_for_on_enabled_fire_count(times).await.unwrap();
-}
-
-#[then(regex = "^`(.*)`'s (audio|(?:device|display) video) local Track is \
-                (muted|unmuted)$")]
-async fn then_local_track_mute_state(
-    world: &mut World,
-    id: String,
-    kind: String,
-    muted: String,
-) {
-    let member = world.get_member(&id).unwrap();
-    let (media_kind, source_kind) = parse_media_kinds(&kind).unwrap();
-    let track = member
-        .room()
-        .local_tracks()
-        .await
-        .unwrap()
-        .get_track(media_kind, source_kind)
-        .await
-        .unwrap();
-    let muted = muted.as_str() == "muted";
-    assert_eq!(muted, track.muted().await.unwrap());
-}
-
-#[then(regex = "^`(.*)`'s (audio|(?:device|display) video) local Track is \
-                stopped$")]
-async fn then_track_is_stopped(world: &mut World, id: String, kind: String) {
-    let member = world.get_member(&id).unwrap();
-    let (media_kind, source_kind) = parse_media_kinds(&kind).unwrap();
-    let is_stopped = member
-        .room()
-        .local_tracks()
-        .await
-        .unwrap()
-        .get_track(media_kind, source_kind)
-        .await
-        .unwrap()
-        .free_and_check()
-        .await
-        .unwrap();
-    assert!(is_stopped);
-}
